Always build for default architecture
The architecture we build for is defined in the compiler activation script. For
linux-64, this is Nocona.

The LLAMA_NATIVE environment variable will make the build system optimize for
the architecture of the build platform, making it incompatible with anything
earlier.

This is recorded in a cbc.yaml option to make this recorded in the build
string, because there will be other variants here later, optimizing the package
for different archs. (Although, we should increment the build number and the
version will most probably be incremented too, so this isn't strictly needed.)